reorganized the repository
Deleted the old repository and reogranized it under a new one.  There's a directory for each language or platform specific code base.  I took a Udemy class in one language and I'm porting the example game that was built in the class to multiple languages. The goal being to main consistency between the versions.